+++ src/main/java/duke/commands/ListCommand.java
@@ -1,0 +1,19 @@
+package duke.commands;
+
+import duke.exceptions.DukeException;
+import duke.storage.Storage;
+import duke.tasks.TaskList;
+import duke.views.UI;
+
+public class ListCommand extends Command {
+
+    @Override
+    public void execute(UI ui, TaskList tasks, Storage storage) throws DukeException {
+        if (tasks.size() == 0) {
+            this.commandStatus = "You currently have 0 tasks added. Add one now!";
+        } else {
+            this.commandStatus = tasks.toString();
+        }
+        ui.printCommandOutput(this);
+    }
+}
